feat(transaction): history transaction

+++ src/main/java/adi_kurniawan/springboot_kash_api/model/transaction/InquiryRequest.java
@@ -6,11 +6,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigInteger;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 @Builder
 public class InquiryRequest {
     @NotNull
-    private Integer accountNumber;
+    private BigInteger accountNumber;
 }
