+++ src/main/java/adi_kurniawan/springboot_kash_api/repository/PocketRepository.java
@@ -5,6 +5,7 @@ import adi_kurniawan.springboot_kash_api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 
@@ -13,7 +14,7 @@ import java.util.Optional;
 public interface PocketRepository extends JpaRepository<Pocket, Integer> {
     Optional<Pocket> findFirstByUserAndAccountNumber(User user, Integer id);
 
-    Optional<Pocket> findFirstByAccountNumber(Integer id);
+    Optional<Pocket> findFirstByAccountNumber(BigInteger id);
 
     List<Pocket> findAllByUserId(Integer userId);
 }
